youtube/requirements.txt
amqp==5.0.6
billiard==3.6.4.0
cached-property==1.5.2
celery==5.0.5
certifi==2021.5.30
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
Django==2.2.6
django-nested-admin==3.3.3
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
idna==2.10
importlib-metadata==4.5.0
kombu==5.1.0
prompt-toolkit==3.0.18
PyJWT==2.1.0
python-monkey-business==1.0.0
pytz==2021.1
redis==3.5.3
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.5
vine==5.0.0
wcwidth==0.2.5
zipp==3.4.1
